pyproject.toml
[project]
name = "media-article-writer"
version = "0.1.0"
description = "CLI tool for generating human-like media articles using AI"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "google-generativeai>=0.3.0",
    "requests>=2.31.0",
    "exa-py>=1.0.0",
    "google-api-python-client>=2.100.0",
    "rich>=13.0.0",
    "click>=8.1.0",
    "crewai>=0.80.0",
    "crewai-tools>=0.15.0",
    "google-genai>=1.57.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.1",
    "responses>=0.23.0",
    "freezegun>=1.2.0",
]

[project.scripts]
article-writer = "src.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
only-include = ["src"]
